# api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "直播互动引擎 - API 网关服务",
        "title": "Live Interact Engine - API Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/v1/danmaku/send": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "在房间内发送弹幕消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Danmaku"
                ],
                "summary": "发送弹幕",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "发送弹幕请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mapper.SendDanmakuReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "弹幕信息",
                        "schema": {
                            "$ref": "#/definitions/mapper.DanmakuResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/danmaku/subscribe": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "通过 SSE 实时订阅房间内的弹幕消息",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Danmaku"
                ],
                "summary": "订阅房间弹幕",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "房间ID",
                        "name": "room_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实时弹幕流",
                        "schema": {
                            "$ref": "#/definitions/mapper.DanmakuResp"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/gift/list": {
            "get": {
                "description": "获取平台全量礼物列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gifts"
                ],
                "summary": "获取礼物列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mapper.ListGiftsResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/gift/send": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "用户向主播发送礼物，扣除用户钱包余额，生成礼物赠送记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gifts"
                ],
                "summary": "发送礼物",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mapper.SendGiftReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mapper.GiftRecordResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/leaderboard/{room_id}": {
            "get": {
                "description": "获取指定房间的礼物赠送排行榜 (按累计送礼金额降序)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leaderboard"
                ],
                "summary": "获取房间排行榜",
                "parameters": [
                    {
                        "type": "string",
                        "description": "房间 ID (UUID)",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "返回前 N 名 (默认 100, 最大 1000)",
                        "name": "top_n",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mapper.LeaderboardResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v1/wallet/{user_id}/balance": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取指定用户的钱包余额（通常获取自己的钱包）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "获取钱包余额",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户 ID (UUID)",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mapper.WalletResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "mapper.DanmakuResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "mentioned_user_id": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "mapper.GiftRecordResp": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "anchor_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "gift_id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "status": {
                    "description": "pending, success, failed",
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "mapper.GiftResp": {
            "type": "object",
            "properties": {
                "cache_key": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "icon_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "status": {
                    "description": "online, offline, limited_time",
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "vip_only": {
                    "type": "boolean"
                }
            }
        },
...
        "mapper.LeaderboardEntryResp": {
            "type": "object",
            "properties": {
                "rank": {
                    "type": "integer"
                },
                "score": {
                    "description": "累计送礼金额",
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "mapper.LeaderboardResp": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mapper.LeaderboardEntryResp"
                    }
                },
                "room_id": {
                    "type": "string"
                }
            }
        },
        "mapper.ListGiftsResp": {
            "type": "object",
            "properties": {
                "gifts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mapper.GiftResp"
                    }
                }
            }
        },
...
        "mapper.SendDanmakuReq": {
            "type": "object",
            "required": [
                "content",
                "room_id",
                "type",
                "user_id",
                "username"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "mentioned_user_id": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "type": {
                    "type": "integer",
                    "maximum": 3,
                    "minimum": 0
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "mapper.SendGiftReq": {
            "type": "object",
            "required": [
                "amount",
                "anchor_id",
                "gift_id",
                "idempotency_key",
                "room_id",
                "user_id"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "anchor_id": {
                    "type": "string"
                },
                "gift_id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "mapper.WalletResp": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "version_number": {
                    "description": "乐观锁版本号",
                    "type": "integer"
                }
            }
        }
    }
}